More core functionality, fixes. (Apparently, the NES is 8-bit)

+++ MNES.Core/Machine/Mappers/Implementation/M000.cs
@@ -14,7 +14,7 @@ namespace MNES.Core.Machine.Mappers.Implementation
             new(MapperBank.AccessorType.CPU, MapperBank.BankType.PRG_RAM, new Range(0x6000, 0x8000), new(0x0000, 0x2000))
         };
 
-        public override ushort this[ushort index] { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public override byte this[ushort index] { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public M000(MachineState machine) : base(machine) { }
